pyproject.toml
[project]
name = "djangoproject"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-urls (>=1.1.4,<2.0.0)",
    "django-shortcuts (>=1.6,<2.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-filter (>=25.1,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
